backend/requirements.txt
# pip install -r requirements.txt


Django==5.1.2

django-environ==0.11.2

django-split-settings==1.3.2

djangorestframework==3.15.2

djangorestframework-camel-case==1.4.2

drf-yasg==1.21.8

django-cors-headers==4.6.0

django-extensions==4.1

# python-dotnet~=1.0.1
pythonnet==3.0.3

requests~=2.32.3

django-cors-headers==4.6.0

channels==4.2.2
 